Issue #3056: Review archive operations permissions - cleanup for permissions manager
