+++ api/src/main/java/com/epam/pipeline/manager/security/storage/StoragePermissionManager.java
@@ -24,6 +24,7 @@ import com.epam.pipeline.entity.datastorage.DataStorageWithShareMount;
 import com.epam.pipeline.entity.datastorage.NFSStorageMountStatus;
 import com.epam.pipeline.entity.datastorage.nfs.NFSDataStorage;
 import com.epam.pipeline.entity.security.acl.AclClass;
+import com.epam.pipeline.entity.user.DefaultRoles;
 import com.epam.pipeline.manager.EntityManager;
 import com.epam.pipeline.manager.datastorage.DataStoragePathLoader;
 import com.epam.pipeline.manager.quota.QuotaService;
@@ -35,6 +36,7 @@ import com.epam.pipeline.security.acl.AclPermission;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.ListUtils;
+import org.springframework.security.acls.domain.GrantedAuthoritySid;
 import org.springframework.security.acls.model.Sid;
 import org.springframework.stereotype.Service;
 
@@ -136,6 +138,20 @@ public class StoragePermissionManager {
             storages.clear();
             storages.addAll(filtered);
         }
+    }
+
+    public boolean storageArchiveReadPermissions(final AbstractDataStorage storage) {
+        return grantPermissionManager.isOwnerOrAdmin(storage.getOwner())
+                || permissionHelper.isAllowed(READ, storage) && checkStorageArchiveRoles();
+    }
+
+    private boolean checkStorageArchiveRoles() {
+        final GrantedAuthoritySid archiveManager = new GrantedAuthoritySid(
+                DefaultRoles.ROLE_STORAGE_ARCHIVE_MANAGER.getName());
+        final GrantedAuthoritySid archiveReader = new GrantedAuthoritySid(
+                DefaultRoles.ROLE_STORAGE_ARCHIVE_READER.getName());
+        return permissionHelper.getSids().stream()
+                .anyMatch(sid -> sid.equals(archiveManager) || sid.equals(archiveReader));
     }
 
     private boolean checkPermissions(final List<AclPermission> permissions,
